+++ projects/angular-toastify/src/lib/toast.ts
@@ -1,4 +1,5 @@
 import { ToastType } from './toast-type';
+import { Observable, Subject } from 'rxjs';
 
 let toastIdentitySequence = 0;
 
@@ -8,10 +9,14 @@ export class Toast {
     type: ToastType;
     readonly time: number;
 
+    $resetToast: Subject<any>;
+
     constructor(message: string, type: ToastType) {
         this.message = message;
         this.type = type;
         this.time = new Date().getTime();
         this.id = toastIdentitySequence++;
+
+        this.$resetToast = new Subject();
     }
 }