+++ projects/angular-toastify/src/lib/toastify-toast/toastify-toast.component.ts
@@ -1,17 +1,32 @@
 import {
-  ChangeDetectionStrategy, ChangeDetectorRef, Component, EventEmitter, HostListener, Input, NgZone, OnDestroy, OnInit, Output
-} from '@angular/core';
-import { ToastType } from '../toast-type';
-import { Toast } from '../toast';
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  ElementRef,
+  EventEmitter,
+  HostListener,
+  Input,
+  NgZone,
+  OnDestroy,
+  OnInit,
+  Output,
+  ViewChild,
+} from "@angular/core";
+import { ToastType } from "../toast-type";
+import { Toast } from "../toast";
+import { interval } from "rxjs";
+import { throttle, throttleTime } from "rxjs/operators";
 
 @Component({
   // tslint:disable-next-line:component-selector
-  selector: 'lib-toastify-toast',
-  templateUrl: './toastify-toast.component.html',
-  styleUrls: ['./toastify-toast.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  selector: "lib-toastify-toast",
+  templateUrl: "./toastify-toast.component.html",
+  styleUrls: ["./toastify-toast.component.scss"],
+  // changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ToastifyToastComponent implements OnInit, OnDestroy {
+  @ViewChild("progressBar") progressBar: ElementRef<HTMLElement>;
+  @ViewChild("progressBarCover") progressBarCover: ElementRef<HTMLElement>;
 
   @Input() autoClose = 5000;
   @Input() hideProgressBar = false;
@@ -19,33 +34,70 @@ export class ToastifyToastComponent implements OnInit, OnDestroy {
   @Input() pauseOnVisibilityChange = true;
   @Input() closeOnClick = true;
   @Input() toast: Toast;
-  @Input() iconLibrary: 'material' | 'font-awesome' | 'none';
+  @Input() iconLibrary: "material" | "font-awesome" | "none";
 
   @Output() dismissEvent = new EventEmitter();
 
+  private expectedAutoDismissTime: number;
   private autoDismissTimeout: any;
   private autoCloseRemaining: number;
 
   private pauseTime: number;
   private startTime: number;
 
+  private _progressBarAnimation: number;
+  private _$updateTimer;
+
   ToastType = ToastType;
   running = false;
 
-  constructor(private _cd: ChangeDetectorRef, private _zone: NgZone) {
-  }
+  constructor(private _cd: ChangeDetectorRef, private _zone: NgZone) {}
 
   ngOnInit(): void {
     this.autoCloseRemaining = this.autoClose;
     this.startTime = this.toast.time;
+    this.toast.$resetToast.subscribe(() => this.resetToastTimer());
     // Do not start timer when toast is prompted while window is out of focus
-    if (this.handleVisibilityChange && document.visibilityState === 'visible') {
+    if (this.handleVisibilityChange && document.visibilityState === "visible") {
       this.startCloseTimer();
     }
+
+    // Start progress bar animation
+    this.triggerProgressBarAnimation();
+  }
+
+  private triggerProgressBarAnimation(): void {
+    // Cancel previous animlation to avoid leaks
+    if (this._progressBarAnimation !== undefined) {
+      cancelAnimationFrame(this._progressBarAnimation);
+    }
+
+    // Start animation
+    const frame = () => {
+      if (this.running) {
+        const remainingTime = Math.max(0, this.expectedAutoDismissTime - new Date().getTime());
+        const percentage = 100 - ((remainingTime / this.autoClose) * 100);
+        this.progressBarCover.nativeElement.style.width = percentage + "%";
+        if (percentage <= 0) return;
+      }
+      this._progressBarAnimation = requestAnimationFrame(frame);
+    };
+    this._progressBarAnimation = requestAnimationFrame(frame);
   }
 
   ngOnDestroy(): void {
+    // Cancel animation
+    if (this._progressBarAnimation) {
+      cancelAnimationFrame(this._progressBarAnimation);
+      this._progressBarAnimation = undefined;
+    }
+    // Clear auto close timeout
     this.clearTimerTimeout();
+    // Complete all observables
+    this.toast.$resetToast?.complete();
+    this.toast.$resetToast = null;
+    this._$updateTimer?.complete();
+    this._$updateTimer = null;
   }
 
   startCloseTimer(): void {
@@ -53,15 +105,17 @@ export class ToastifyToastComponent implements OnInit, OnDestroy {
       return;
     }
 
-    this.startTime = new Date().getTime();
     this.running = true;
+    this.expectedAutoDismissTime =
+      new Date().getTime() + this.autoCloseRemaining;
     this.autoDismissTimeout = this._zone.runOutsideAngular(() =>
       setTimeout(() => {
         this._zone.run(() => {
           this.dismissEvent.emit();
           this._cd.markForCheck();
         });
-      }, this.autoCloseRemaining));
+      }, this.autoCloseRemaining)
+    );
   }
 
   pauseCloseTimer(): void {
@@ -71,11 +125,20 @@ export class ToastifyToastComponent implements OnInit, OnDestroy {
     // Calculate the elapsed time, subtract remaining time
     this.pauseTime = new Date().getTime();
     const elapsed = this.pauseTime - this.startTime;
-    this.autoCloseRemaining -= elapsed;
+    this.autoCloseRemaining = this.autoClose - elapsed;
+  }
+
+  resetToastTimer() {
+    this.clearTimerTimeout();
+    this.running = false;
+    this.startTime = new Date().getTime();
+    this.autoCloseRemaining = this.autoClose;
+    this.startCloseTimer();
   }
 
   clearTimerTimeout(): void {
     if (this.autoDismissTimeout !== undefined) {
+      this.expectedAutoDismissTime = undefined;
       clearTimeout(this.autoDismissTimeout);
     }
   }
@@ -89,7 +152,7 @@ export class ToastifyToastComponent implements OnInit, OnDestroy {
     this.dismissEvent.emit();
   }
 
-  @HostListener('click')
+  @HostListener("click")
   handleHostClick(): void {
     if (this.closeOnClick) {
       this.clearTimerTimeout();
@@ -97,27 +160,28 @@ export class ToastifyToastComponent implements OnInit, OnDestroy {
     }
   }
 
-  @HostListener('mouseenter')
+  @HostListener("mouseenter")
   handleMouseEnter(): void {
     if (this.pauseOnHover) {
       this.pauseCloseTimer();
     }
   }
 
-  @HostListener('mouseleave')
+  @HostListener("mouseleave")
   handleMouseLeave(): void {
     if (this.pauseOnHover) {
       this.startCloseTimer();
+      this.startTime = new Date().getTime() + (this.startTime - this.pauseTime);
     }
   }
 
-  @HostListener('document:visibilitychange')
+  @HostListener("document:visibilitychange")
   handleVisibilityChange(): void {
     if (!this.pauseOnVisibilityChange) {
       return;
     }
 
-    if (document.visibilityState !== 'visible') {
+    if (document.visibilityState !== "visible") {
       this.pauseCloseTimer();
       this._cd.detectChanges();
     } else {