+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ##### 1.0.5
 
-- Added a new property `preventDuplicates` on the Toastify Container. When set to true, toasts with identical texts will no longer create duplicates (duplicates are removed). Credits go to [@JoranLive](https://github.com/JoranLive) for implementing this feature.
+- Added a new property `preventDuplicates` on the Toastify Container. When set to true, toasts with identical texts will no longer create duplicates (duplicates are reset). Credits go to [@JoranLive](https://github.com/JoranLive) for implementing this feature.
 
 ##### 1.0.4