+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+##### 1.0.5
+
+- Added a new property `preventDuplicates` on the Toastify Container. When set to true, toasts with identical texts will no longer create duplicates (duplicates are reset). Credits go to [@JoranLive](https://github.com/JoranLive) for implementing this feature.
+
 ##### 1.0.4
 
 - Module no longer imports `BrowserModule`. Angular Toastify can now be used in submodules.